feat(agent-task): add advanced settings tab
